+++ docs/api/ext.md
@@ -1,7 +1,5 @@
 # `pydantic_ai.ext`
 
-::: pydantic_ai.ext.temporal
-
 ::: pydantic_ai.ext.langchain
 
 ::: pydantic_ai.ext.aci
